requirements.txt
torch>=2.2.1
transformers>=4.38.1
apache-beam>=2.50.0
notebook>=7.0.7
scikit-learn>=1.4.0
google-cloud-aiplatform>=1.38.1 